split the navigate btn
