+++ src/components/entered/FormRegisterComponent.tsx
@@ -13,19 +13,6 @@ const FormregisterComponent = ({ title, handleClick }) => {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
 
-  const navigate = useNavigate();
-
-  const handleClickNavigate = () => {
-    navigate('/auth-new-user');
-  };
-  const handleNavigateForSignUp = () => {
-    navigate('/register-new-user');
-  };
-
-  const handleAttach = () => {
-    navigate('/attach-files');
-  };
-
   localStorage.setItem('ERW', email);
   return (
     <>
@@ -51,20 +38,9 @@ const FormregisterComponent = ({ title, handleClick }) => {
         />
         <FormEnteredFormButton>
           <ButtonElement
-            variant="outlined"
-            text="Sign-UP"
-            handleClick={handleNavigateForSignUp}
-          />
-
-          <ButtonElement
-            variant="outlined"
-            text="Sign-In"
-            handleClick={handleClickNavigate}
-          />
-          <ButtonElement
-            variant="outlined"
-            text="Attach"
-            handleClick={handleAttach}
+            variant="text"
+            text={title}
+            handleClick={() => handleClick(email, password)}
           />
         </FormEnteredFormButton>
       </FormEnteredForm>
