+++ src/components/navigate/index.ts
@@ -1,0 +1,2 @@
+import NavigateWrapper from './NavigateWrapper';
+export default NavigateWrapper;
